mas: Check validity of instruments/samples used in patterns
When a pattern uses an invalid instrument it should be deleted from the
resulting MAS file. This patch flags all valid instruments as valid so that
a function can check all patterns after the whole module is loaded.

Note that IT has two modes: use samples or use instruments. If the sample
mode is selected, only samples with a non-zero length are considered valid.

In MOD and S3M only samples with non-zero length are valid instruments.

The old code used to check for this error in the MOD importer has been
removed and turned into a function common for all formats.