mas: Sanitize modules with too many pattern orders
Some module files have a pattern order array that is bigger than the limit
of the MAS format (200). However, the additional pattern orders may simply
be empty (maybe the tool likes to save 256 pattern orders always).

This patch adds a check to ensure that any pattern order after index 200 is
empty. If not, it will show a warning to the user.

In any case, if the pattern order size is over 200, it is always clamped to
200 entries.